backend/requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AWS
boto3>=1.34.72
botocore>=1.34.72

# LangChain/LangGraph
langchain==0.1.0
langgraph==0.0.20
langchain-aws==0.1.0

# Document Processing
pypdf==3.17.0
python-docx==1.1.0
markdown==3.5.0
beautifulsoup4==4.12.0

# Vector/Embeddings
opensearch-py==2.4.0
numpy==1.26.0

# API
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Utils
requests==2.31.0
tenacity==8.2.3

# Development
pytest==7.4.0
pytest-asyncio==0.21.0
black==24.1.0
ruff==0.1.0
